Add functionality for sharing folders
The links that are created expire after a certain amount of time, and
for accuracy the date of expiry and comparing of dates is done in the
UTC timezone for maximum accuracy. Then the UTC time is transformed into
the users locale(the Date constructor has a default timezone of UTC
when nothing is specified, so it's better to store it like that instead
of having inaccuracies in comparisons because of PostgreSQL using the
local timezone as default for now()).